api/requirements.txt
# FastAPI and server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
pgvector>=0.2.4

# Face recognition
face-recognition>=1.3.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.1.0
dlib>=19.24.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Optional production dependencies (uncomment as needed):
# retinaface>=0.0.16          # For RetinaFace detection (faster, more accurate)
# mediapipe>=0.10.0           # For MediaPipe/BlazeFace detection (mobile-friendly)
# insightface>=0.7.3          # For ArcFace/InsightFace embeddings (512D, state-of-the-art)
# onnxruntime>=1.15.0         # For ONNX model inference
# onnxruntime-gpu>=1.15.0     # For GPU-accelerated ONNX inference

